cmdeploy: show DNS info at begin and end of cmdeploy run
